Complete the sign up form
